@@ -13,17 +13,38 @@ apply plugin: 'kotlin-android'
 apply plugin: 'kotlin-android-extensions'
 apply plugin: 'org.jetbrains.dokka-android'
 apply plugin: 'com.novoda.bintray-release'
+apply plugin: 'maven-publish'
+apply plugin: "com.jfrog.artifactory"
+apply plugin: 'checkstyle'
+preBuild.dependsOn 'Checkstyle'
+preBuild.dependsOn 'detekt'
+
+checkstyle {
+    ignoreFailures false
+    showViolations true
+    toolVersion = "8.11"
+}
+
+task Checkstyle(type: Checkstyle, group: "codeCorrectness") {
+    configFile rootProject.file('checkstyle.xml')
+    source 'src/'
+    ignoreFailures false
+    showViolations true
+    include '**/*.java'
+    include '**/*.kt'
+
+    classpath = files()
+}
 
 android {
-    compileSdkVersion 26
+    compileSdkVersion 28
     defaultConfig {
         minSdkVersion 21
-        targetSdkVersion 26
+        targetSdkVersion 28
         versionCode 1
-        versionName "1.0"
+        versionName "1.0.0"
 
         testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
-
     }
 
     buildTypes {
@@ -35,15 +56,56 @@ android {
 
 }
 
+publishing {
+    publications {
+        aar(MavenPublication) {
+            setGroupId("com.hushed")
+            setArtifactId("waveformview")
+            version android.defaultConfig.versionName
+            artifact("$buildDir/outputs/aar/waveformview-release.aar")
+        }
+    }
+}
+
+artifactory {
+    contextUrl = "${artifactory_contextUrl}"   //The base Artifactory URL if not overridden by the publisher/resolver
+    publish {
+        repository {
+            repoKey = 'gradle-dev-local'
+            username = "${artifactory_user}"
+            password = "${artifactory_password}"
+            maven = true
+        } defaults {
+            // Tell the Artifactory Plugin which artifacts should be published to Artifactory.
+            publications('aar')
+            publishArtifacts = true
+            // Properties to be attached to the published artifacts.
+            properties = ['qa.level': 'basic', 'dev.team': 'core']
+            // Publish generated POM files to Artifactory (true by default)
+            publishPom = true
+        }
+    }
+    resolve {
+        repository {
+            repoKey = 'gradle-dev'
+            username = "${artifactory_user}"
+            password = "${artifactory_password}"
+            maven = true
+        }
+    }
+}
+
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
 
-    implementation 'com.android.support:appcompat-v7:26.1.0'
+    implementation 'androidx.appcompat:appcompat:1.0.2'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
     testImplementation 'junit:junit:4.12'
-    androidTestImplementation 'com.android.support.test:runner:1.0.1'
-    androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.1'
+    androidTestImplementation 'androidx.test:runner:1.1.1'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.1.1'
     compile "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
-    compile 'org.jetbrains.kotlinx:kotlinx-coroutines-core:0.22.1'
+    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-core:1.2.1'
+    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-android:1.2.1'
 }
 
 repositories {
@@ -67,15 +129,6 @@ task dokkaMarkdown(type: org.jetbrains.dokka.gradle.DokkaTask) {
         prefix = "android" // will match kotlin.internal and all sub-packages of it
         suppress = true
     }
-}
-
-publish {
-    userOrg = 'siy1121'
-    groupId = 'space.siy'
-    artifactId = 'waveformview'
-    publishVersion = '1.0.0'
-    desc = 'Provide the view to show audio wave form'
-    website = 'https://github.com/SIY1121/WaveFormViewDemo'
 }
 
 tasks.withType(Javadoc).all {
